requirements.txt
anyascii==0.3.2
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-appconf==1.0.5
django-check-seo==0.6.2
django-classy-tags==4.0.0
django-cms==3.11.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filer==2.2.4
django-filter==22.1
django-formtools==2.4
django-js-asset==2.0.0
django-meta==2.2.0
django-modelcluster==6.0
django-mptt==0.14.0
django-permissionedforms==0.1
django-polymorphic==3.1.0
django-sekizai==4.0.0
django-taggit==3.1.0
django-templated-mail==1.1.1
django-treebeard==4.7
django-user-accounts==3.2.0
djangocms-admin-style==3.2.3
djangocms-page-meta==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
draftjs-exporter==2.1.7
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
httpie==3.2.1
idna==3.4
l18n==2021.3
lxml==4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.4
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
svglib==1.5.1
telepath==0.3
tinycss2==1.2.1
tzdata==2023.3
Unidecode==1.1.2
urllib3==1.26.15
wagtail==4.2.2
webencodings==0.5.1
Willow==1.4.1
zope.event==4.6
zope.interface==6.0
